backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
gunicorn
django-storages
boto3
python-dotenv
Pillow
cloudinary
django-cloudinary-storage
django-cities-light
django-countries